pyproject.toml
[project]
name = "surah-splitter"
version = "0.1.0"
description = "A tool to split Surah audio files into individual Ayahs."
keywords = [
    "islamic",
    "islam",
    "quran",
    "surah",
    "ayah",
    "audio",
    "splitter",
    "audio-processing",
]
license = { file = "LICENSE" }
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "OdyAsh", email = "ashicsresources@gmail.com" }
]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Below 2 libraries are only required if you'll run .ipynb
    #   or .py files as notebooks using "# %%" VSCode cells.
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    # Core dependencies
    "cyclopts>=3.17.0",
    "loguru>=0.7.3",
    "pydub>=0.25.1",
    "rich>=14.0.0",
    "whisperx>=3.4.2",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    # Azure OpenAI integration
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
    # Real-time audio processing
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    # Web interface
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "jinja2>=3.1.0",
    # Audio processing enhancements
    "webrtcvad>=2.0.10",
    "scipy>=1.10.0",
    # Machine learning
    "scikit-learn>=1.3.0",
    # Streaming STT
    "azure-cognitiveservices-speech>=1.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.org/simple/"
